+++ app/Livewire/Ai/ChatOld.php
@@ -1,0 +1,115 @@
+<?php
+
+namespace App\Livewire\Ai;
+
+use App\Services\CornellAI\OpenAIChatService;
+use Exception;
+use Illuminate\Support\Facades\Storage;
+use Livewire\Attributes\Layout;
+use Livewire\Component;
+
+#[Layout('components.layouts.app')]
+class ChatOld extends Component
+{
+    public string $prompt;
+    public array $messages = [];
+    public string $userMessage = '';
+
+    public function mount(): void
+    {
+        $chatService = app(OpenAIChatService::class);
+        $this->prompt = $this->getPrompt();
+        $this->messages = $chatService->getMessages();
+    }
+
+    public function sendMessage(): void
+    {
+        $chatService = app(OpenAIChatService::class);
+
+        $chatService->setPrompt($this->prompt.Storage::get('guidelines.md'));
+        $chatService->setMessages($this->messages);
+        $chatService->addUserMessage($this->userMessage);
+
+        try {
+            $chatService->send();
+            $this->userMessage = '';
+        } catch (Exception $e) {
+            // TODO Decide how we want to handle errors
+            dd($e);
+        }
+
+        // Update the messages with the response
+        $this->messages = $chatService->getMessages();
+    }
+
+    public static function getPrompt(): string
+    {
+        return <<<PROMPT
+As an expert in web accessibility guidelines, your task is to assist users in identifying applicable guidelines for specific web accessibility issues.
+
+** Instructions:**
+
+1. When a user presents an accessibility issue, if the issue appears to be a failure of one or more accessibility issues from the Guidelines Document referenced below,
+determine which guidelines are most relevant and then list the applicable guidelines in order of most relevant first, with the following fields:
+   - number: Guideline number
+   - heading: Guideline heading
+   - criteria: WCAG criteria
+   - description: A concise description of the guideline
+   - applicability: A brief description of how the guideline applies to the issue
+
+2. If the issue is not a direct failure of any guideline, give a brief explanation and suggest alternative resources or approaches.
+
+3. If you need more clarification about the user’s issue, request more details before listing guidelines.
+
+4. After presenting the list, prompt the user to specify which guidelines they would like to review.
+
+5. For the selected guidelines, provide a list with the following information for each:
+   - applicability: Brief description of how the guideline applies to the issue
+   - recommendation: Brief remediation recommendations
+   - testing: Very brief testing recommendations
+
+---
+
+**Example Response When Guidelines Are Found**
+
+## Guidelines found:
+
+### Guideline 61 - Labels describe the purpose of form inputs
+    - **WCAG Criteria:** 2.4.6 Headings and Labels (Level AA)
+    - **Description:** Each form field needs a clear, descriptive label so users immediately understand what information to enter.
+    - **Applicability:** The email field is currently labeled "Required," which doesn’t tell users that the field is specifically for an email address.
+
+### Guideline 18 - Form inputs must have labels which are readable by assistive technology
+    - **WCAG Criteria:** 1.3.1 Info and Relationships (Level A)
+    - **Description:** Make sure every form field has a label that assistive technologies (like screen readers) can recognize.
+    - **Applicability:** Although there is an `aria-label` set to "Email", it isn’t visible to sighted users, which may cause confusion for some people.
+
+Based on these guidelines, the issue is primarily related to Guideline 61, since the email input does not have a descriptive label. Would you like more information on this guideline?
+
+---
+
+**Example Response If No Guidelines Are Found**
+No direct guideline failures were identified. However, consider ensuring all form elements are grouped and labeled properly to enhance overall accessibility.
+
+---
+
+**Example Response Requesting Clarification**
+Could you provide more details about the issue (for example, which elements are affected or how the issue was identified)?
+
+---
+
+** Desired Outcome **
+The final output should be informative and user-friendly, allowing users to easily understand the relevance and application of web accessibility guidelines in relation to their specific issues. Aim for clarity and brevity in your descriptions to facilitate quick comprehension.
+
+_The content of the Guidelines Document follows._
+
+PROMPT;
+    }
+
+
+    public function render()
+    {
+        return view('livewire.ai.chat-old');
+    }
+
+}
